requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
cachetools==5.5.2
    # via google-auth
certifi==2025.7.9
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
fastapi==0.116.1
    # via dialogflow-website-n-google-chat
google-api-core==2.25.1
    # via
    #   google-api-python-client
    #   google-cloud-dialogflow
google-api-python-client==2.176.0
    # via dialogflow-website-n-google-chat
google-auth==2.40.3
    # via
    #   dialogflow-website-n-google-chat
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-dialogflow
google-auth-httplib2==0.2.0
    # via
    #   dialogflow-website-n-google-chat
    #   google-api-python-client
google-auth-oauthlib==1.2.2
    # via dialogflow-website-n-google-chat
google-cloud-dialogflow==2.41.2
    # via dialogflow-website-n-google-chat
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.73.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.73.1
    # via google-api-core
h11==0.16.0
    # via uvicorn
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.6
    # via dialogflow-website-n-google-chat
markupsafe==3.0.2
    # via jinja2
oauthlib==3.3.1
    # via requests-oauthlib
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-dialogflow
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-dialogflow
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.3
    # via httplib2
python-dotenv==1.1.1
    # via dialogflow-website-n-google-chat
python-multipart==0.0.20
    # via dialogflow-website-n-google-chat
requests==2.32.4
    # via
    #   dialogflow-website-n-google-chat
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.47.1
    # via fastapi
typing-extensions==4.14.1
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via dialogflow-website-n-google-chat
